// qto-ui/src/app/models/service-view.model.ts
import { AbstractBaseModel } from "./abstract-base-model";

export class ServiceView extends AbstractBaseModel {
  locationId: number;
  clientLocationId: number;
  orderId: number;
  address: string;
  address1: string;
  address2: string;
  city: string;
  stateProvince: string;
  postalCode: string;
  status: string;
  provisioner: string;
  i90ProjectManager: string;
  projectManager: string;
  provider: string;
  customerRequestedInstall: Date;
  siteSurveyDue: Date;
  siteSurveySubmit: Date;
  providerOrderSubmitted: Date;
  networkProviderFoc: Date;
  dataProvisioningComplete: Date;
  greatestMilestoneName: string;
  greatestMilestoneDate: Date;
  clientLocationType: string;
  clientLocationInfo: string;
  companyName: string;
  companyId: number;
  endCustomerClientId: string;
  parentCompanyName: string;
  parentCompanyId: number;
  parentCompanyClientId: string;
  speed: string;
  type: string;
  openJeop: string;
  openJeopResponsibilities: string;
  latestNote: string;
  vertekProjectManager: number;
  orderType: string;
  created: Date;
  clientServiceId: string;
  statusAge: number;
  providerCircuitId: string;
  mrc: number;
  nrc: number;
  lconPhone: string;
  summaryBill: number;
  activeInactive: string;
  contractSignedDate: Date;
  contractTerm: string;
  circuitTermEndDate: Date;
  inventoryAddedDate: Date;
  macdCount: number;
  countOpenDisputes: number;
  icbFlag: string;
  mrcDisputed: number;
  nrcDisputed: number;
  disputeTypes: string;
  subOrderType: string;
  accountNumber: string;
  serviceBilledTo: string;
  followUpDate: Date | null;

  selected?: boolean;
  progressPercentage: number;
}